requirements.txt
torch==2.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
xlsxwriter==3.2.0
filterpy==1.4.5
matplotlib==3.8.2
seaborn==0.13.0
tensorboard==2.15.1
pyyaml==6.0.1
